requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements-dev.txt requirements-dev.in
#
appdirs==1.4.2            # via setuptools
autoflake==0.6.6
autopep8==1.2.4
click==6.7                # via pip-tools
coverage==4.3.4           # via pytest-cov
decorator==4.0.11         # via ipython, traitlets
first==2.0.1              # via pip-tools
flake8==3.3.0
ipython-genutils==0.1.0   # via traitlets
ipython==5.3.0
isort==4.2.5
mccabe==0.6.1             # via flake8
packaging==16.8           # via setuptools
pep8-naming==0.4.1
pep8==1.7.0               # via autopep8
pexpect==4.2.1            # via ipython
pickleshare==0.7.4        # via ipython
pip-tools==1.8.0
prompt-toolkit==1.0.13    # via ipython
ptyprocess==0.5.1         # via pexpect
py==1.4.32                # via pytest
pycodestyle==2.3.1        # via flake8
pyflakes==1.5.0           # via autoflake, flake8
pygments==2.2.0           # via ipython
pyparsing==2.1.10         # via packaging
pytest-cov==2.4.0
pytest-django==3.1.2
pytest==3.0.6
simplegeneric==0.8.1      # via ipython
six==1.10.0               # via packaging, pip-tools, prompt-toolkit, setuptools, traitlets
traitlets==4.3.2          # via ipython
wcwidth==0.1.7            # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools                # via ipython, pytest
